requirements.txt
dashscope
tree-sitter>=0.20.0
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
tree-sitter-java
tree-sitter-cpp
tree-sitter-c
tree-sitter-go
tree-sitter-html
tree-sitter-css

numpy
scikit-learn
cryptography
gitpython
watchdog
click
rich
pydantic
aiofiles
asyncio-throttle
fastapi
uvicorn[standard]
python-multipart
